Convert/requirements.txt
pillow>=10.0.0
pillow-heif>=0.13.0